Domain as part of the UserAgent
This allows you to search for specific accesses from a domain in the server log files. I find helpful for troubleshooting, but also as an indication of where the query comes to the plugin.
The data protection does, I think, no harm.
From the home network 'localhost' would be transmitted, as no secret, otherwise the domain would be transmitted in addition to the public IP when the server sets it.
Who runs its own server (Dedicated Server or Virtual Server) at which one can resolve the host over the IP even now.

Only with shared hosting there is the advantage of additionally finding the domain in the log.